move  bot
