--- core/bot/__init__.py
@@ -1,1 +1,0 @@
-from .bot import UniversitasTerbukaBot
